test(api): rewrite telegram upload init bdd tests
